backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
web3
python-dotenv
requests
pydantic
slowapi
google-genai
